scraped titles
Ran the youtube_video_scraper and youtube_video_titles scripts to generate my list of titles, which includes over 63,000 videos. Not included are the individual channel video lists, because these will be unique to the channels you choose. The titles.csv and titles.txt files are included here for documentation purposes.